pyproject.toml
[tool.poetry]
name = "beautiful-sites"
version = "0.1.0"
description = ""
authors = ["CatCoderr <catcoderr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2"
django-rest-framework = "^0.1.0"
django-cors-headers = "^3.14.0"
dj-database-url = "^2.0.0"
mysqlclient = "^2.1.1"
gunicorn = "20.1.0"
django-health-check = "^3.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

